fix: timeout and race issues in tests
This commit fixes a few issues with timeouts and race issues
encountered in the CI, based on a debugging session with
Joice and me.

The major changes are concerned with reducing the timeouts by
half, and removing random from TestGrace to make it deterministic. We
have also removed the SIGINT notify listener since it conflicted
with the test timeout. Also, moved the port for TestGrace to be
greater than 10000.